change the project structure.
