+++ src/Neo.Compiler.CSharp/MethodConvert/Helpers/StackHelpers.cs
