+++ src/Neo.Compiler.CSharp/MethodConvert/MethodConvert.cs
@@ -38,7 +38,15 @@ namespace Neo.Compiler
         private CallingConvention _callingConvention = CallingConvention.Cdecl;
         private bool _inline;
         private bool _internalInline;
-        private bool _initslot;
+        // _initSlot is a boolean flag that determines whether an INITSLOT instruction
+        // should be added at the beginning of the method's bytecode.
+        // It's used to allocate space for local variables and parameters in the Neo VM.
+        // The _initSlot flag allows the compiler to avoid adding unnecessary INITSLOT
+        // instructions for methods that don't need local variables or parameters.
+        // It's typically set to false for inline methods or external method calls.
+        // By using this flag, the compiler can efficiently manage stack space allocation
+        // for method execution in the Neo VM, only allocating space when necessary.
+        private bool _initSlot;
         private readonly Dictionary<IParameterSymbol, byte> _parameters = new(SymbolEqualityComparer.Default);
         private readonly List<(ILocalSymbol, byte)> _variableSymbols = new();
         private readonly Dictionary<ILocalSymbol, byte> _localVariables = new(SymbolEqualityComparer.Default);
@@ -83,169 +91,114 @@ namespace Neo.Compiler
             this._context = context;
             this._checkedStack.Push(context.Options.Checked);
         }
-
-        #endregion
-
-        #region Variables
-
-        private byte AddLocalVariable(ILocalSymbol symbol)
-        {
-            byte index = (byte)(_localVariables.Count + _anonymousVariables.Count);
-            _variableSymbols.Add((symbol, index));
-            _localVariables.Add(symbol, index);
-            if (_localsCount < index + 1)
-                _localsCount = index + 1;
-            _blockSymbols.Peek().Add(symbol);
-            return index;
-        }
-
-        private byte AddAnonymousVariable()
-        {
-            byte index = (byte)(_localVariables.Count + _anonymousVariables.Count);
-            _anonymousVariables.Add(index);
-            if (_localsCount < index + 1)
-                _localsCount = index + 1;
-            return index;
-        }
-
-        private void RemoveAnonymousVariable(byte index)
-        {
-            if (_context.Options.Optimize.HasFlag(CompilationOptions.OptimizationType.Basic))
-                _anonymousVariables.Remove(index);
-        }
-
-        private void RemoveLocalVariable(ILocalSymbol symbol)
-        {
-            if (_context.Options.Optimize.HasFlag(CompilationOptions.OptimizationType.Basic))
-                _localVariables.Remove(symbol);
-        }
-
-        #endregion
-
-        #region Instructions
-        private Instruction AddInstruction(Instruction instruction)
-        {
-            _instructions.Add(instruction);
-            return instruction;
-        }
-
-        private Instruction AddInstruction(OpCode opcode)
-        {
-            return AddInstruction(new Instruction
-            {
-                OpCode = opcode
-            });
-        }
-
-        private SequencePointInserter InsertSequencePoint(SyntaxNodeOrToken? syntax)
-        {
-            return new SequencePointInserter(_instructions, syntax);
-        }
-
-        private SequencePointInserter InsertSequencePoint(SyntaxReference? syntax)
-        {
-            return new SequencePointInserter(_instructions, syntax);
-        }
-
-        private SequencePointInserter InsertSequencePoint(Location? location)
-        {
-            return new SequencePointInserter(_instructions, location);
-        }
-
         #endregion
 
         #region Convert
+
+        /// <summary>
+        /// This method is responsible for converting a method into a Neo VM bytecode.
+        /// </summary>
+        /// <param name="model">The semantic model of the method</param>
         public void Convert(SemanticModel model)
         {
+            // Step 1: Determine if the method is extern or empty
+            // This checks if:
+            // a) The method is marked as extern, e.g.:
+            //    public static extern int ExternMethod(int arg);
+            // b) The containing type (class) of the method has no declaring syntax references
+            //    This can happen with partial classes or classes defined in other assemblies
             if (Symbol.IsExtern || Symbol.ContainingType.DeclaringSyntaxReferences.IsEmpty)
             {
+                // Step 2a: Handle extern or empty methods
+                // This path is taken for:
+                // - Methods explicitly marked as extern (see example above)
+                // - Methods in classes defined in other assemblies or partial classes, e.g.:
+                //   public partial class MyClass
+                //   {
+                //       // This method might be defined in another file or assembly
+                //       public void EmptyMethod();
+                //   }
                 if (Symbol.Name == "_initialize")
                 {
+                    // Special handling for _initialize method
+                    // This method is used for contract initialization
                     ProcessStaticFields(model);
-                    if (_context.StaticFieldCount > 0)
-                    {
-                        _instructions.Insert(0, new Instruction
-                        {
-                            OpCode = OpCode.INITSSLOT,
-                            Operand = new[] { (byte)_context.StaticFieldCount }
-                        });
-                    }
+                    InsertStaticFieldInitialization();
                 }
                 else
                 {
+                    // Convert extern method
+                    // This usually involves creating a stub or placeholder for the external implementation
                     ConvertExtern();
                 }
             }
             else
             {
+                // Step 2b: Handle regular methods
+                // This path is taken for normal methods with implementations in the current compilation unit
+                // These methods require full processing and conversion
+                // Example of a regular method:
+                // public class MyClass
+                // {
+                //     public int RegularMethod(int arg)
+                //     {
+                //         return arg * 2;
+                //     }
+                // }
+                // Set syntax node if available
                 if (!Symbol.DeclaringSyntaxReferences.IsEmpty)
                     SyntaxNode = Symbol.DeclaringSyntaxReferences[0].GetSyntax();
-                switch (Symbol.MethodKind)
-                {
-                    case MethodKind.Constructor:
-                        ProcessFields(model);
-                        ProcessConstructorInitializer(model);
-                        break;
-                    case MethodKind.StaticConstructor:
-                        ProcessStaticFields(model);
-                        break;
-                    default:
-                        if (Symbol.Name.StartsWith("_") && !Helper.IsInternalCoreMethod(Symbol))
-                            throw new CompilationException(Symbol, DiagnosticId.InvalidMethodName, $"The method name {Symbol.Name} is not valid.");
-                        break;
-                }
+
+                // Step 3: Process method based on its kind
+                // This handles special cases for constructors and static constructors
+                // Examples:
+                // public MyClass() { } // Constructor
+                // static MyClass() { } // Static constructor
+                InitializeFieldsBasedOnMethodKind(model);
+
+                // Step 4: Validate method name
+                // Ensures that method names starting with '_' are valid
+                // Example of an invalid method name:
+                // public void _invalidMethod() { } // This would throw an exception
+                ValidateMethodName();
+
+                // Step 5: Convert modifiers
+                // process ModifierAttribute
                 var modifiers = ConvertModifier(model).ToArray();
+
+                // Step 6: Convert the main method body
+                // This is where the actual method implementation is converted
                 ConvertSource(model);
-                if (Symbol.MethodKind == MethodKind.StaticConstructor && _context.StaticFieldCount > 0)
-                {
-                    _instructions.Insert(0, new Instruction
-                    {
-                        OpCode = OpCode.INITSSLOT,
-                        Operand = new[] { (byte)_context.StaticFieldCount }
-                    });
-                }
-                if (_initslot)
-                {
-                    byte pc = (byte)_parameters.Count;
-                    byte lc = (byte)_localsCount;
-                    if (IsInstanceMethod(Symbol)) pc++;
-                    if (pc > 0 || lc > 0)
-                    {
-                        _instructions.Insert(0, new Instruction
-                        {
-                            OpCode = OpCode.INITSLOT,
-                            Operand = new[] { lc, pc }
-                        });
-                    }
-                }
-                foreach (var (fieldIndex, attribute) in modifiers)
-                {
-                    var disposeInstruction = ExitModifier(model, fieldIndex, attribute);
-                    if (disposeInstruction is not null && _returnTarget.Instruction is null)
-                    {
-                        _returnTarget.Instruction = disposeInstruction;
-                    }
-                }
+
+                // Step 7: Insert initialization instructions if needed
+                // This includes initializing static fields and local variables
+                // This might insert INITSLOT instruction to the beginning of the method
+                // Example:
+                // static int staticField = 10;
+                // public void Method()
+                // {
+                //     int localVar = 5;
+                //     // ... rest of the method
+                // }
+                InsertInitializationInstructions();
+
+                // Step 8: Process modifiers (exit)
+                // Handle any cleanup or additional instructions required by modifiers
+                ProcessModifiersExit(model, modifiers);
             }
-            if (_returnTarget.Instruction is null)
-            {
-                if (_instructions.Count > 0 && _instructions[^1].OpCode == OpCode.NOP && _instructions[^1].SourceLocation is not null)
-                {
-                    _instructions[^1].OpCode = OpCode.RET;
-                    _returnTarget.Instruction = _instructions[^1];
-                }
-                else
-                {
-                    _returnTarget.Instruction = AddInstruction(OpCode.RET);
-                }
-            }
-            else
-            {
-                // it comes from modifier clean up
-                AddInstruction(OpCode.RET);
-            }
+
+            // Step 9: Finalize the method
+            // Add RET instruction to the end of the method
+            // This ensures proper method termination and return instruction
+            FinalizeMethod();
+
+            // Step 10: Optimize the instructions if needed
+            // Basic optimization to remove unnecessary NOP instructions
             if (_context.Options.Optimize.HasFlag(CompilationOptions.OptimizationType.Basic))
                 BasicOptimizer.RemoveNops(_instructions);
+
+            // Step 11: Set the start target
+            // Mark the first instruction as the entry point of the method
             _startTarget.Instruction = _instructions[0];
         }
 
@@ -269,8 +222,8 @@ namespace Neo.Compiler
                 SyntaxNode syntaxNode;
                 if (field.DeclaringSyntaxReferences.IsEmpty)
                 {
-                    // Have to process the string.Empty specially since it has no AssociatedSymbol
-                    // thus will return directly without this if check.
+                    // Special handling for string.Empty is required as it lacks an AssociatedSymbol.
+                    // Without this check, the method would return prematurely, bypassing necessary processing.
                     if (field.ContainingType.ToString() == "string" && field.Name == "Empty")
                     {
                         preInitialize?.Invoke();
@@ -348,393 +301,7 @@ namespace Neo.Compiler
             }
         }
 
-        private IEnumerable<(byte fieldIndex, AttributeData attribute)> ConvertModifier(SemanticModel model)
-        {
-            foreach (var attribute in Symbol.GetAttributesWithInherited())
-            {
-                if (attribute.AttributeClass?.IsSubclassOf(nameof(ModifierAttribute)) != true)
-                    continue;
 
-                JumpTarget notNullTarget = new();
-                byte fieldIndex = _context.AddAnonymousStaticField();
-                AccessSlot(OpCode.LDSFLD, fieldIndex);
-                AddInstruction(OpCode.ISNULL);
-                Jump(OpCode.JMPIFNOT_L, notNullTarget);
-
-                MethodConvert constructor = _context.ConvertMethod(model, attribute.AttributeConstructor!);
-                CreateObject(model, attribute.AttributeClass, null);
-                foreach (var arg in attribute.ConstructorArguments.Reverse())
-                    Push(arg.Value);
-                Push(attribute.ConstructorArguments.Length);
-                AddInstruction(OpCode.PICK);
-                EmitCall(constructor);
-                AccessSlot(OpCode.STSFLD, fieldIndex);
-
-                notNullTarget.Instruction = AccessSlot(OpCode.LDSFLD, fieldIndex);
-                var enterSymbol = attribute.AttributeClass.GetAllMembers()
-                    .OfType<IMethodSymbol>()
-                    .First(p => p.Name == nameof(ModifierAttribute.Enter) && p.Parameters.Length == 0);
-                MethodConvert enterMethod = _context.ConvertMethod(model, enterSymbol);
-                EmitCall(enterMethod);
-                yield return (fieldIndex, attribute);
-            }
-        }
-
-        private Instruction? ExitModifier(SemanticModel model, byte fieldIndex, AttributeData attribute)
-        {
-            var exitSymbol = attribute.AttributeClass!.GetAllMembers()
-                .OfType<IMethodSymbol>()
-                .First(p => p.Name == nameof(ModifierAttribute.Exit) && p.Parameters.Length == 0);
-            MethodConvert exitMethod = _context.ConvertMethod(model, exitSymbol);
-            if (exitMethod.IsEmpty) return null;
-            var instruction = AccessSlot(OpCode.LDSFLD, fieldIndex);
-            EmitCall(exitMethod);
-            return instruction;
-        }
-        #endregion
-
-        #region Helper
-
-        /// <summary>
-        /// load parameter value
-        /// </summary>
-        /// <param name="parameter"></param>
-        /// <returns></returns>
-        private Instruction LdArgSlot(IParameterSymbol parameter)
-        {
-            if (_context.TryGetCaptruedStaticField(parameter, out var staticFieldIndex))
-            {
-                //using created static fields
-                return AccessSlot(OpCode.LDSFLD, staticFieldIndex);
-            }
-            if (Symbol.MethodKind == MethodKind.AnonymousFunction && !_parameters.ContainsKey(parameter))
-            {
-                //create static fields from captrued parameter
-                var staticIndex = _context.GetOrAddCapturedStaticField(parameter);
-                CapturedLocalSymbols.Add(parameter);
-                return AccessSlot(OpCode.LDSFLD, staticIndex);
-            }
-            // local parameter in current method
-            byte index = _parameters[parameter];
-            return AccessSlot(OpCode.LDARG, index);
-        }
-
-        /// <summary>
-        /// store value to parameter
-        /// </summary>
-        /// <param name="parameter"></param>
-        /// <returns></returns>
-        private Instruction StArgSlot(IParameterSymbol parameter)
-        {
-            if (_context.TryGetCaptruedStaticField(parameter, out var staticFieldIndex))
-            {
-                //using created static fields
-                return AccessSlot(OpCode.STSFLD, staticFieldIndex);
-            }
-            if (Symbol.MethodKind == MethodKind.AnonymousFunction && !_parameters.ContainsKey(parameter))
-            {
-                //create static fields from captrued parameter
-                var staticIndex = _context.GetOrAddCapturedStaticField(parameter);
-                CapturedLocalSymbols.Add(parameter);
-                return AccessSlot(OpCode.STSFLD, staticIndex);
-            }
-            // local parameter in current method
-            byte index = _parameters[parameter];
-            return AccessSlot(OpCode.STARG, index);
-        }
-
-        /// <summary>
-        /// load local variable value
-        /// </summary>
-        /// <param name="local"></param>
-        /// <returns></returns>
-        private Instruction LdLocSlot(ILocalSymbol local)
-        {
-            if (_context.TryGetCaptruedStaticField(local, out var staticFieldIndex))
-            {
-                //using created static fields
-                return AccessSlot(OpCode.LDSFLD, staticFieldIndex);
-            }
-            if (Symbol.MethodKind == MethodKind.AnonymousFunction && !_localVariables.ContainsKey(local))
-            {
-                //create static fields from captrued local
-                byte staticIndex = _context.GetOrAddCapturedStaticField(local);
-                CapturedLocalSymbols.Add(local);
-                return AccessSlot(OpCode.LDSFLD, staticIndex);
-            }
-            // local variables in current method
-            byte index = _localVariables[local];
-            return AccessSlot(OpCode.LDLOC, index);
-        }
-
-        /// <summary>
-        /// store value to local variable
-        /// </summary>
-        /// <param name="local"></param>
-        /// <returns></returns>
-        private Instruction StLocSlot(ILocalSymbol local)
-        {
-            if (_context.TryGetCaptruedStaticField(local, out var staticFieldIndex))
-            {
-                //using created static fields
-                return AccessSlot(OpCode.STSFLD, staticFieldIndex);
-            }
-            if (Symbol.MethodKind == MethodKind.AnonymousFunction && !_localVariables.ContainsKey(local))
-            {
-                //create static fields from captrued local
-                byte staticIndex = _context.GetOrAddCapturedStaticField(local);
-                CapturedLocalSymbols.Add(local);
-                return AccessSlot(OpCode.STSFLD, staticIndex);
-            }
-            byte index = _localVariables[local];
-            return AccessSlot(OpCode.STLOC, index);
-        }
-
-        private Instruction AccessSlot(OpCode opcode, byte index)
-        {
-            return index >= 7
-                ? AddInstruction(new Instruction { OpCode = opcode, Operand = new[] { index } })
-                : AddInstruction(opcode - 7 + index);
-        }
-
-        private bool TryProcessInlineMethods(SemanticModel model, IMethodSymbol symbol, IReadOnlyList<SyntaxNode>? arguments)
-        {
-            SyntaxNode? syntaxNode = null;
-            if (!symbol.DeclaringSyntaxReferences.IsEmpty)
-                syntaxNode = symbol.DeclaringSyntaxReferences[0].GetSyntax();
-
-            if (syntaxNode is not BaseMethodDeclarationSyntax syntax) return false;
-            if (!symbol.GetAttributesWithInherited().Any(attribute => attribute.ConstructorArguments.Length > 0
-                    && attribute.AttributeClass?.Name == nameof(MethodImplAttribute)
-                    && attribute.ConstructorArguments[0].Value is not null
-                    && (MethodImplOptions)attribute.ConstructorArguments[0].Value! == MethodImplOptions.AggressiveInlining))
-                return false;
-
-            _internalInline = true;
-
-            using (InsertSequencePoint(syntax))
-            {
-                if (arguments is not null) PrepareArgumentsForMethod(model, symbol, arguments, CallingConvention.Cdecl);
-                if (syntax.Body != null) ConvertStatement(model, syntax.Body);
-            }
-            return true;
-        }
-
-        private void PrepareArgumentsForMethod(SemanticModel model, IMethodSymbol symbol, IReadOnlyList<SyntaxNode> arguments, CallingConvention callingConvention = CallingConvention.Cdecl)
-        {
-            var namedArguments = arguments.OfType<ArgumentSyntax>().Where(p => p.NameColon is not null).Select(p => (Symbol: (IParameterSymbol)model.GetSymbolInfo(p.NameColon!.Name).Symbol!, p.Expression)).ToDictionary(p => p.Symbol, p => p.Expression, (IEqualityComparer<IParameterSymbol>)SymbolEqualityComparer.Default);
-            IEnumerable<IParameterSymbol> parameters = symbol.Parameters;
-            if (callingConvention == CallingConvention.Cdecl)
-                parameters = parameters.Reverse();
-            foreach (IParameterSymbol parameter in parameters)
-            {
-                if (namedArguments.TryGetValue(parameter, out ExpressionSyntax? expression))
-                {
-                    ConvertExpression(model, expression);
-                }
-                else if (parameter.IsParams)
-                {
-                    if (arguments.Count > parameter.Ordinal)
-                    {
-                        if (arguments.Count == parameter.Ordinal + 1)
-                        {
-                            expression = arguments[parameter.Ordinal] switch
-                            {
-                                ArgumentSyntax argument => argument.Expression,
-                                ExpressionSyntax exp => exp,
-                                _ => throw new CompilationException(arguments[parameter.Ordinal], DiagnosticId.SyntaxNotSupported, $"Unsupported argument: {arguments[parameter.Ordinal]}"),
-                            };
-                            Conversion conversion = model.ClassifyConversion(expression, parameter.Type);
-                            if (conversion.Exists)
-                            {
-                                ConvertExpression(model, expression);
-                                continue;
-                            }
-                        }
-                        for (int i = arguments.Count - 1; i >= parameter.Ordinal; i--)
-                        {
-                            expression = arguments[i] switch
-                            {
-                                ArgumentSyntax argument => argument.Expression,
-                                ExpressionSyntax exp => exp,
-                                _ => throw new CompilationException(arguments[i], DiagnosticId.SyntaxNotSupported, $"Unsupported argument: {arguments[i]}"),
-                            };
-                            ConvertExpression(model, expression);
-                        }
-                        Push(arguments.Count - parameter.Ordinal);
-                        AddInstruction(OpCode.PACK);
-                    }
-                    else
-                    {
-                        AddInstruction(OpCode.NEWARRAY0);
-                    }
-                }
-                else
-                {
-                    if (arguments.Count > parameter.Ordinal)
-                    {
-                        switch (arguments[parameter.Ordinal])
-                        {
-                            case ArgumentSyntax argument:
-                                if (argument.NameColon is null)
-                                {
-                                    ConvertExpression(model, argument.Expression);
-                                    continue;
-                                }
-                                break;
-                            case ExpressionSyntax ex:
-                                ConvertExpression(model, ex);
-                                continue;
-                            default:
-                                throw new CompilationException(arguments[parameter.Ordinal], DiagnosticId.SyntaxNotSupported, $"Unsupported argument: {arguments[parameter.Ordinal]}");
-                        }
-                    }
-                    Push(parameter.ExplicitDefaultValue);
-                }
-            }
-        }
-
-        private Instruction IsType(VM.Types.StackItemType type)
-        {
-            return AddInstruction(new Instruction
-            {
-                OpCode = OpCode.ISTYPE,
-                Operand = new[] { (byte)type }
-            });
-        }
-
-        private Instruction ChangeType(VM.Types.StackItemType type)
-        {
-            return AddInstruction(new Instruction
-            {
-                OpCode = OpCode.CONVERT,
-                Operand = new[] { (byte)type }
-            });
-        }
-
-        private void InitializeFieldForObject(SemanticModel model, IFieldSymbol field, InitializerExpressionSyntax? initializer)
-        {
-            ExpressionSyntax? expression = null;
-            if (initializer is not null)
-            {
-                foreach (ExpressionSyntax e in initializer.Expressions)
-                {
-                    if (e is not AssignmentExpressionSyntax ae)
-                        throw new CompilationException(initializer, DiagnosticId.SyntaxNotSupported, $"Unsupported initializer: {initializer}");
-                    if (SymbolEqualityComparer.Default.Equals(field, model.GetSymbolInfo(ae.Left).Symbol))
-                    {
-                        expression = ae.Right;
-                        break;
-                    }
-                }
-            }
-            if (expression is null)
-                PushDefault(field.Type);
-            else
-                ConvertExpression(model, expression);
-        }
-
-        private void CreateObject(SemanticModel model, ITypeSymbol type, InitializerExpressionSyntax? initializer)
-        {
-            ISymbol[] members = type.GetAllMembers().Where(p => !p.IsStatic).ToArray();
-            IFieldSymbol[] fields = members.OfType<IFieldSymbol>().ToArray();
-            if (fields.Length == 0 || type.IsValueType || type.IsRecord)
-            {
-                AddInstruction(type.IsValueType || type.IsRecord ? OpCode.NEWSTRUCT0 : OpCode.NEWARRAY0);
-                foreach (IFieldSymbol field in fields)
-                {
-                    AddInstruction(OpCode.DUP);
-                    InitializeFieldForObject(model, field, initializer);
-                    AddInstruction(OpCode.APPEND);
-                }
-            }
-            else
-            {
-                for (int i = fields.Length - 1; i >= 0; i--)
-                    InitializeFieldForObject(model, fields[i], initializer);
-                Push(fields.Length);
-                AddInstruction(OpCode.PACK);
-            }
-            IMethodSymbol[] virtualMethods = members.OfType<IMethodSymbol>().Where(p => p.IsVirtualMethod()).ToArray();
-            if (!type.IsRecord && virtualMethods.Length > 0)
-            {
-                byte index = _context.AddVTable(type);
-                AddInstruction(OpCode.DUP);
-                AccessSlot(OpCode.LDSFLD, index);
-                AddInstruction(OpCode.APPEND);
-            }
-        }
-
-        private Instruction Jump(OpCode opcode, JumpTarget target)
-        {
-            return AddInstruction(new Instruction
-            {
-                OpCode = opcode,
-                Target = target
-            });
-        }
-
-        /// <summary>
-        /// Convert a throw expression or throw statement to OpCodes.
-        /// </summary>
-        /// <param name="model">The semantic model providing context and information about the Throw.</param>
-        /// <param name="exception">The content of exception</param>
-        /// <exception cref="CompilationException">Only a single parameter is supported for exceptions.</exception>
-        /// <example>
-        /// throw statement:
-        /// <code>
-        /// if (shapeAmount <= 0)
-        /// {
-        ///     throw new Exception("Amount of shapes must be positive.");
-        /// }
-        ///</code>
-        /// throw expression:
-        /// <code>
-        /// string a = null;
-        /// var b = a ?? throw new Exception();
-        /// </code>
-        /// <code>
-        /// var first = args.Length >= 1 ? args[0] : throw new Exception();
-        /// </code>
-        /// </example>
-        /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-throw-expression">The throw expression</seealso>
-        /// <seealso href="https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/statements/exception-handling-statements#the-try-catch-statement">Exception-handling statements - throw</seealso>
-        private void Throw(SemanticModel model, ExpressionSyntax? exception)
-        {
-            if (exception is not null)
-            {
-                ITypeSymbol type = model.GetTypeInfo(exception).Type!;
-                if (type.IsSubclassOf(nameof(scfx::Neo.SmartContract.Framework.UncatchableException), includeThisClass: true))
-                {
-                    AddInstruction(OpCode.ABORT);
-                    return;
-                }
-            }
-            switch (exception)
-            {
-                case ObjectCreationExpressionSyntax expression:
-                    switch (expression.ArgumentList?.Arguments.Count)
-                    {
-                        case null:
-                        case 0:
-                            Push("exception");
-                            break;
-                        case 1:
-                            ConvertExpression(model, expression.ArgumentList.Arguments[0].Expression);
-                            break;
-                        default:
-                            throw new CompilationException(expression, DiagnosticId.MultiplyThrows, "Only a single parameter is supported for exceptions.");
-                    }
-                    break;
-                case null:
-                    AccessSlot(OpCode.LDLOC, _exceptionStack.Peek());
-                    break;
-                default:
-                    ConvertExpression(model, exception);
-                    break;
-            }
-            AddInstruction(OpCode.THROW);
-        }
         #endregion
     }
